[defn-tags] Move tags normalization logic
