+++ python_modules/dagster/dagster/_core/definitions/assets.py
@@ -71,12 +71,12 @@ from dagster._core.definitions.utils import (
     DEFAULT_IO_MANAGER_KEY,
     normalize_group_name,
     validate_asset_owner,
-    validate_tags_strict,
 )
 from dagster._core.errors import DagsterInvalidDefinitionError, DagsterInvariantViolationError
 from dagster._utils import IHasInternalInit
 from dagster._utils.merger import merge_dicts
 from dagster._utils.security import non_secure_md5_hash_str
+from dagster._utils.tags import validate_tags_strict
 from dagster._utils.warnings import ExperimentalWarning, disable_dagster_warnings
 
 if TYPE_CHECKING:
