+++ python_modules/dagster/dagster/_core/definitions/decorators/source_asset_decorator.py
@@ -20,7 +20,7 @@ from dagster._core.definitions.partition import PartitionsDefinition
 from dagster._core.definitions.resource_annotation import get_resource_args
 from dagster._core.definitions.resource_definition import ResourceDefinition
 from dagster._core.definitions.source_asset import SourceAsset, SourceAssetObserveFunction
-from dagster._core.definitions.utils import validate_tags_strict
+from dagster._utils.tags import validate_tags_strict
 from dagster._utils.warnings import disable_dagster_warnings
