+++ python_modules/libraries/dagster-airflow/dagster_airflow/dagster_job_factory.py
@@ -8,8 +8,8 @@ from dagster import (
     ResourceDefinition,
     _check as check,
 )
-from dagster._core.definitions.utils import normalize_tags
 from dagster._core.instance import IS_AIRFLOW_INGEST_PIPELINE_STR
+from dagster._utils.tags import normalize_tags
 
 from dagster_airflow.airflow_dag_converter import get_graph_definition_args
 from dagster_airflow.resources import (
