+++ python_modules/libraries/dagster-dbt/dagster_dbt/dagster_dbt_translator.py
@@ -10,7 +10,7 @@ from dagster import (
     _check as check,
 )
 from dagster._annotations import experimental, public
-from dagster._core.definitions.utils import is_valid_definition_tag_key
+from dagster._utils.tags import is_valid_definition_tag_key
 
 from dagster_dbt.asset_utils import (
     default_asset_key_fn,
